fix factory pattern

+++ app/src/main/java/com/example/tapmouse/settingsActivity/size/factory/SetSize.java
@@ -2,6 +2,6 @@ package com.example.tapmouse.settingsActivity.size.factory;
 
 public interface SetSize {
 
-    double mouseWidth(double width);
-    double mouseHeight(double height);
+    int mouseWidth(int width);
+    int mouseHeight(int height);
 }
